requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
openai>=1.12.0
pydantic>=2.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
passlib>=1.7.4
bcrypt>=4.0.0
python-jose>=3.3.0
email-validator